backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

pymongo==4.6.1
dnspython==2.4.2

groq==0.9.0

python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0
pytest==7.4.0